2.Django/requirements.txt
Django==1.11.10
djangorestframework==3.7.7
